refactor: create runner/collect package
