+++ pkg/runner/step.go
@@ -7,6 +7,7 @@ import (
 	"github.com/kyverno/chainsaw/pkg/apis/v1alpha1"
 	"github.com/kyverno/chainsaw/pkg/client"
 	"github.com/kyverno/chainsaw/pkg/resource"
+	"github.com/kyverno/chainsaw/pkg/runner/collect"
 	"github.com/kyverno/chainsaw/pkg/runner/logging"
 	"github.com/kyverno/chainsaw/pkg/runner/operations"
 	"github.com/kyverno/kyverno/ext/output/color"
@@ -38,7 +39,7 @@ func executeStep(t *testing.T, logger logging.Logger, ctx Context, basePath stri
 		if t.Failed() {
 			t.Cleanup(func() {
 				for _, handler := range step.Spec.OnFailure {
-					collectors, err := collect(handler.Collect)
+					collectors, err := collect.Commands(handler.Collect)
 					if err != nil {
 						logger.Log("COLLEC", color.BoldRed, err)
 						t.Fail()
